[refactor][AssetCondition] Move AssetCondition-related files to a new subfolder
