+++ python_modules/dagster/dagster/_core/definitions/asset_daemon_context.py
@@ -34,8 +34,8 @@ from dagster._core.instance import DynamicPartitionsStore
 
 from ... import PartitionKeyRange
 from ..storage.tags import ASSET_PARTITION_RANGE_END_TAG, ASSET_PARTITION_RANGE_START_TAG
-from .asset_condition import AssetConditionEvaluation, AssetConditionEvaluationState
-from .asset_condition_evaluation_context import (
+from .asset_condition.asset_condition import AssetConditionEvaluation, AssetConditionEvaluationState
+from .asset_condition.asset_condition_evaluation_context import (
     AssetConditionEvaluationContext,
 )
 from .asset_daemon_cursor import AssetDaemonCursor
