+++ python_modules/dagster/dagster/_core/definitions/auto_materialize_rule_evaluation.py
@@ -33,12 +33,11 @@ from dagster._serdes.serdes import (
 from .partition import PartitionsDefinition, SerializedPartitionsSubset
 
 if TYPE_CHECKING:
-    from dagster._core.definitions.asset_condition import AssetSubsetWithMetadata
-
-    from .asset_condition import (
+    from .asset_condition.asset_condition import (
         AssetConditionEvaluation,
         AssetConditionEvaluationWithRunIds,
         AssetConditionSnapshot,
+        AssetSubsetWithMetadata,
     )
 
 
@@ -148,7 +147,7 @@ def deserialize_auto_materialize_asset_evaluation_to_asset_condition_evaluation_
     """Provides a backcompat layer to allow deserializing old AutoMaterializeAssetEvaluation
     objects into the new AssetConditionEvaluationWithRunIds objects.
     """
-    from .asset_condition import AssetConditionEvaluationWithRunIds
+    from .asset_condition.asset_condition import AssetConditionEvaluationWithRunIds
 
     class BackcompatDeserializer(BackcompatAutoMaterializeAssetEvaluationSerializer):
         @property
@@ -213,7 +212,7 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(NamedTupleSerializer):
     def _asset_condition_snapshot_from_rule_snapshot(
         self, rule_snapshot: AutoMaterializeRuleSnapshot
     ) -> "AssetConditionSnapshot":
-        from .asset_condition import AssetConditionSnapshot, RuleCondition
+        from .asset_condition.asset_condition import AssetConditionSnapshot, RuleCondition
 
         unique_id_parts = [rule_snapshot.class_name, rule_snapshot.description]
         unique_id = hashlib.md5("".join(unique_id_parts).encode()).hexdigest()
@@ -233,7 +232,7 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(NamedTupleSerializer):
         is_partitioned: bool,
         rule_snapshot: AutoMaterializeRuleSnapshot,
     ) -> "AssetConditionEvaluation":
-        from .asset_condition import (
+        from .asset_condition.asset_condition import (
             AssetConditionEvaluation,
             AssetSubsetWithMetadata,
             HistoricalAllPartitionsSubsetSentinel,
@@ -278,7 +277,7 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(NamedTupleSerializer):
         is_partitioned: bool,
         decision_type: AutoMaterializeDecisionType,
     ) -> Optional["AssetConditionEvaluation"]:
-        from .asset_condition import (
+        from .asset_condition.asset_condition import (
             AssetConditionEvaluation,
             AssetConditionSnapshot,
             HistoricalAllPartitionsSubsetSentinel,
@@ -371,7 +370,7 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(NamedTupleSerializer):
         whitelist_map: WhitelistMap,
         context: UnpackContext,
     ) -> "AssetConditionEvaluationWithRunIds":
-        from .asset_condition import (
+        from .asset_condition.asset_condition import (
             AndAssetCondition,
             AssetConditionEvaluation,
             AssetConditionSnapshot,
