+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/updated_scenarios/custom_condition_scenarios.py
@@ -1,5 +1,5 @@
 from dagster import AutoMaterializeRule
-from dagster._core.definitions.asset_condition import RuleCondition
+from dagster._core.definitions.asset_condition.asset_condition import RuleCondition
 from dagster._core.definitions.auto_materialize_policy import AutoMaterializePolicy
 
 from ..asset_daemon_scenario import (
